refactor burndatanftmodal from walletdatanftmx
